fix(debug): correct flag adding for debug
The way it was set up before would always add the debug flag, this fixes
it and makes it only apply the flag if I've typed "make debug". The proper
way to compile with debug mode is "make clean debug" for the first time,
and then just "make debug" subsequent times. To go back to normal, all you
need is a "make re".