chore(select): move file, little changes
Rename the file from select.c to ft_select.c to reflect the function
name's change, and also change the name of the function call in the
main. I've switched the main's printing to an ft_printf, and added a
newline there. I've also changed my proof-of-concept ft_select to return
the first argument, instead of a hard-coded string with a newline.