feat(debug): show function in debug message
Update print_debug function to take an additional argument for the name
of the function that the message is coming from. Also add a macro that
passes in __func__ as the first argument, which is a (char *) for the
name of the current function. Switch all use of print_debug in
srcs/terminal.c to use this macro.