add(config): file for various settings
Adds a config file, which can control the padding between the strings in
select mode, the character for the padding on the output string (as well
as the amount of characters to pad the output strings with), the string
used when trimming strings that are too long in select mode (as well as
it's length), and the overall minimum length that a string can be before
select decides that it won't draw anything. By changing these, you can
radically change the behavior of ft_select. I would like to eventually
include in here various colors, as well as a parameter to force a
certain amount of columns, among other things.