chore(libft): change type of setup function
Change setup function to be a void, instead of an int. This removes the
ability for it to fail for now, but it really doesn't need to as all
it's ever gonna do is set parameters inside the struct. The thing we're
really concerned about failing is the tcsetattr that follows.