refactor(select_info): move file into directory
Move single select_info.c file into a directory. This file was also
split into an init.c, which initializes the struct, and update.c, which
has the functions for updating the terminal's size and figuring out the
new columns. This is because I'll have other things related to
select_info to add, and they needed to be properly organized.