add(info setup): create things for info
Create the strings iter struct out of the input to ft_select. Also adds
a function that goes through the struct and sets a grid index, based on
how many columns we have and the size of the terminal. This function may
end up getting moved to update, I'm not sure yet.