add(update function): get max string length
This function isn't used yet, but it needed a home. I'll be using it
when deleting strings from the list, if the string that was deleted is
the longest. It needs tweaked to be like the others, but since I'm not
actually using it yet I'm not going to do that.