chore(restore): change error code handling
Remove parameter from select_return, as it's position is changing. It
used to be the "quit from anywhere return button, give it the error code
to pass through" but it never really turned into that, it's more of the
reset to normal button. It doesn't need to take any arguments or
anything, so now it only has to return an error code if something inside
it fails.