refactor(select_loop): make function void
Change function to return void, instead of the string. It was already
setting the string as a parameter, no need to return it. Also null
protect at the beginning, and break in the loop if the list is empty -
this ensures that I quit when the last thing is deleted.