feat(signals): signal handling implemented
Implement signal handling with a signal holding function and a jump
table to signal functions

The main problem I had with a single signal handler function who set an
internal variable, is that if I was in the middle of a read (almost
always), it would stay in the read. That's why the signal_catch function
sends a null character to the input - to skip the read. The read in my
loop ignores null input, so this doesn't cause any problems.

This makes handling quit signals very easy, because I can just have them
quit the program like normal. Stop and continue is also very easy, but
it's more of a "pause" - it's stuck inside that function until it
continues, and I don't actually handle the continue signal.