chore(fourbyte input): jump table instead of if/else
This is an annoying one to squish into a jump table, but the first two
bytes are the same for all of these so I can easily cut that out. An
array of 65535 function pointers isn't the best thing ever, but it's
also hardly anything at all. Plus if this ends up being bad, all I have
to do is reverse this commit, no problem.