norm: single byte input function
I had to cut the brackets on the if statement, add a ternary, and
change the function to send in an int instead of a character, but I
turned 30 some lines into 25. That'll all be ruined when I have to
add the backspace key here soon, but I don't wanna think about that
yet.