fix(linux signals): recreate signal handler in catch function
Add a call to signal in signal_catch to re-establish the signal handler. In
linux, after the signal has been received, first the handler for that signal
is reset to SIG_DFL, and then your handler function is ran. This ensures that
the handler stays activated, for linux (and other unix-based systems). MacOS
and other BSD-based systems don't use this behavior of signal, and this is one
of the reasons that I think we should be learning sigaction for this project.
At the same time though, all that signal handling for interactive applications
is handled by ncurses, so if we're supposed to be using legacy termcaps then
I guess it makes just as much sense to use the old signal interface as well.