feat: don't draw strings if the terminal is too small
Ensures that the columns are set to 0 if the strings won't all fit
on the screen. Adds a check in draw_screen to write a message if your
screen is too small. Also ignores any input while the screen is too
small, which actually causes the bug of not being able to quit. I
haven't yet decided how I'm going to get around that, I didn't really
want to ignore it in all of the functions that deal with input except
the quitting ones, but it seems like that's what has to happen.