feat(quit): quit if screen is too small
Removes the ft_select_quit function, and moves the quit condition to
before the input gets handled. After checking if the input was to quit,
it looks if the screen is being drawn (if there are no columns, then we
can't fit everything on the screen) and if it isn't, then we continue
instead of handling the input.