fix(segfault): deleting the last string fixed
Makes sure that the list has things in it before setting the amount of
columns. Without doing this, the amount of columns would be wrong, which
would cause a seg fault in the function to get the grid position,
because it assumed that everything was fine if I had a number for the
columns, when in reality the columns should have been set to 0 in here.