fix(sigstop): send signal with kill instead of ioctl
As a part of the project, we weren't allowed to use functions other than
a set few. Kill and getpid were not allowed, but ioctl was. The previous
solution for stopping the process in the stop handler was to send in a
^Z with ioctl, but it was extremely inconsistent on linux. This way
works much better, and is actually the proper way to do it.